Resolved comments made by Frank. -Created a Enum list in order to make a list of the leagues in case we need to add more. -Changed value in editTeam where we were receiving int instead of long and that was causing issues -Implemented a method in for validate the league so we can use it and not duplicate code anymore -tested with default profile and populated profile. -Created a method to convert DTo into entity -Tested some endpoints -adjust correctly entities in matchEntity. -Added score into matchEntity -Added logs and exceptions to some methods -Change the endpoint into DTOS. -Added league to Team and Player in order to reuse Services -Created a new Postmanfile

+++ src/main/java/com/iwaconsolti/league/demo/service/MatchService.java
@@ -1,63 +1,86 @@
 package com.iwaconsolti.league.demo.service;
 
 import com.iwaconsolti.league.demo.dto.MatchDTO;
+import com.iwaconsolti.league.demo.dto.PlayerDTO;
 import com.iwaconsolti.league.demo.dto.TeamDTO;
 import com.iwaconsolti.league.demo.entity.MatchEntity;
+import com.iwaconsolti.league.demo.entity.PlayerEntity;
 import com.iwaconsolti.league.demo.entity.TeamEntity;
 import com.iwaconsolti.league.demo.repository.MatchRepository;
 import com.iwaconsolti.league.demo.repository.TeamRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Slf4j
 public class MatchService {
     private final MatchRepository matchRepository;
     private final TeamRepository teamRepository;
+    private final TeamService teamService;
 
-    public MatchService(MatchRepository matchRepository, TeamRepository teamRepository) {
+    public MatchService(MatchRepository matchRepository, TeamRepository teamRepository, TeamService teamService) {
         this.matchRepository = matchRepository;
         this.teamRepository = teamRepository;
+        this.teamService = teamService;
     }
 
-    public void createMatch(TeamDTO team1, TeamDTO team2, String league) {
-        TeamEntity teamONE = teamRepository.findByName(team1.getName());
-        TeamEntity teamTWO = teamRepository.findByName(team2.getName());
+    public void createMatch(TeamDTO teamDTO1, TeamDTO teamDTO2) {
+        TeamEntity teamONE = teamRepository.findByName(teamDTO1.getName());
+        TeamEntity teamTWO = teamRepository.findByName(teamDTO2.getName());
 
-
-        if (teamONE == null && teamTWO == null) {
-            teamONE = new TeamEntity();
-            teamONE.setName(team1.getName());
-            teamONE.setScore(team1.getScore());
-            teamONE.setLeague(league);
-
-            teamRepository.save(teamONE);
-
-            teamTWO = new TeamEntity();
-            teamTWO.setName(team2.getName());
-            teamTWO.setScore(team2.getScore());
-            teamTWO.setLeague(league);
-
-            teamRepository.save(teamTWO);
+        if (teamONE == null) {
+            log.info("Team one doesn't exist.");
+            teamONE = teamService.createTeam(teamDTO1);
         }
+        if(teamTWO == null){
+            log.info("Team one doesn't exist.");
+            teamTWO = teamService.createTeam(teamDTO2);
+        }
+
+        if (!teamDTO1.getLeague().equalsIgnoreCase(teamDTO2.getLeague())) {
+            log.error("Teams have different leagues {} and {}", teamDTO1.getLeague(), teamDTO2.getLeague());
+        }
+
         MatchEntity matchEntity = new MatchEntity(teamONE, teamTWO);
         matchRepository.save(matchEntity);
         log.info("New Match has been created: {} vs {}", teamONE.getName(), teamTWO.getName());
     }
 
     public void deleteAllMatches() {
+        if(matchRepository.count() == 0){
+            log.warn("No matches found to delete");
+            return;
+        }
         matchRepository.deleteAll();
-        log.info("All matches has been deleted");
+        log.info("All matches have been deleted");
     }
 
     public List<MatchDTO> getAllMatches() {
-        return matchRepository.findAll().stream()
+        List<MatchDTO> matches = matchRepository.findAll().stream()
                 .map(match -> new MatchDTO(
-                        new TeamDTO(match.getTeam1().getName(), match.getTeam1().getScore(), match.getTeam1().getID(), match.getTeam1().getLeague()),
-                        new TeamDTO(match.getTeam2().getName(), match.getTeam2().getScore(), match.getTeam2().getID(), match.getTeam2().getLeague())))
+                        new TeamDTO(match.getTeam1().getId(), match.getTeam1().getName(), convertPlayerToDTOList(match.getTeam1().getPlayers()), match.getTeam1().getLeague()),
+                        new TeamDTO(match.getTeam2().getId(), match.getTeam2().getName(), convertPlayerToDTOList(match.getTeam2().getPlayers()), match.getTeam2().getLeague())
+                ))
                 .toList();
+        if (matches.isEmpty()) {
+            log.info("No matches found in this repo");
+        }
+        return matches;
     }
+
+
+    public List<PlayerDTO> convertPlayerToDTOList(List<PlayerEntity> playerEntities) {
+
+        return playerEntities.stream()
+                .map(playerEntity -> new PlayerDTO(
+                        playerEntity.getId(),
+                        playerEntity.getName(),
+                        playerEntity.getTeam().getName())
+                )
+                .collect(Collectors.toList());
+    }
+
 }
